+++ airbnb/src/main/java/com/airbnb/airbnb/auth/MemberAuthenticationEntryPoint.java
@@ -1,13 +1,12 @@
 package com.airbnb.airbnb.auth;
 
+import com.airbnb.airbnb.response.ErrorResponder;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.stereotype.Component;
 
-import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
